Description
This is a whiteboard paint program created in Unity. The scripts were created by using C#. This was a project created within an agile setting in a group of 7 people. I was responsible for the pen tool, the eraser, and various background processes related to those tools. This unfortunately is not the finished product as I do not currently have access to that. But the majority of the functionality within this version was what I was responsible for.